print the first 98 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int i;
-	long long int current = 3, temp, prev = 2;
+	unsigned long int current = 3, temp, prev = 2;
 
 	printf("1, 2, ");
 	for (i = 2; i < 98 ; i++)
